UefiCpuPkg/LocalApicLib: Remove TimerLib dependency
The patch removes API SendInitSipiSipi() and
SendInitSipiSipiAllExcludingSelf() from LocalApicLib, so the library
doesn't depend on TimerLib.
Because many modules in UefiCpuPkg depend on LocalApicLib,  the
change removes the unnecessary TimerLib dependency from most of the
UefiCpuPkg modules.

The patch adds two APIs SendStartupIpi() and
SendStartupIpiAllExcludingSelf() to LocalApicLib, so existing
consumers (MpInitLib and PiSmmCpuDxeSmm/CpuS3.c) of the removed API
can use the new API. The delay originally in LocalApicLib is moved
to the consumer code.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ruiyu Ni <ruiyu.ni@intel.com>
Cc: Jiewen Yao <jiewen.yao@intel.com>
Cc: Eric Dong <eric.dong@intel.com>