UefiCPuPkg/CpuExceptionHandlerLib: Do not depend on SerialPortLib
Today's implementation hard-code dumps the CPU context to serial port
when exception happens.

It's incorrect. Because certain platform may not contain serial port
and it removes the flexibility of dumping CPU context to other debug
channels.

The patch changes to use DEBUG() to dump the CPU context.
DEBUG () can still be enabled with proper value of
PcdDebugProperyMask and PcdFixedDebugPrintErrorLevel in release
build.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ruiyu Ni <ruiyu.ni@intel.com>
Cc: Jiewen Yao <jiewen.yao@intel.com>
Cc: Eric Dong <eric.dong@intel.com>
Cc: Jian J Wang <jian.j.wang@intel.com>